Make scoring tibble

make_scoring_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scoring tibble. The function returns Scoring (a tibble).

Usage

make_scoring_tb(
  name_chr = character(0),
  short_name_chr = character(0),
  domains_unwtd_ls = list(list()),
  scoring_fn_chr = character(0),
  country_chr = character(0),
  dictionary_args_ls = list(list()),
  dictionary_fn_chr = character(0),
  domains_wtd_ls = list(list()),
  dss_ls = list(list()),
  item_prefix_chr = character(0),
  match_chr = character(0),
  scoring_args_ls = list(list()),
  total_unwtd_nm_chr = character(0),
  total_wtd_nm_chr = character(0),
  version_chr = character(0),
  type_chr = "dbl"
)

Arguments

name_chr

Name (a character vector), Default: character(0)

short_name_chr

Short name (a character vector), Default: character(0)

domains_unwtd_ls

Domains unweighted (a list), Default: list(list())

scoring_fn_chr

Scoring function (a character vector), Default: character(0)

country_chr

Country (a character vector), Default: character(0)

dictionary_args_ls

Dictionary arguments (a list), Default: list(list())

dictionary_fn_chr

Dictionary function (a character vector), Default: character(0)

domains_wtd_ls

Domains weighted (a list), Default: list(list())

dss_ls

Datasets (a list), Default: list(list())

item_prefix_chr

Item prefix (a character vector), Default: character(0)

match_chr

Match (a character vector), Default: character(0)

scoring_args_ls

Scoring arguments (a list), Default: list(list())

total_unwtd_nm_chr

Total unweighted name (a character vector), Default: character(0)

total_wtd_nm_chr

Total weighted name (a character vector), Default: character(0)

version_chr

Version (a character vector), Default: character(0)

type_chr

Type (a character vector), Default: 'dbl'

Value

Scoring (a tibble)